requirements.txt
fs
paramiko
pymongo
requests
pyyaml
pycryptodome
faker
click